requirements.txt
Django==1.10.0
djangorestframework==3.4.4
